Rename file to clarify intent
