+++ webapp/components/SegmentIntro.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed } from 'vue'
-import { huc8s } from '~/assets/conus_huc8'
+import { huc8s } from '~/assets/huc8names'
 
 import { useStreamSegmentStore } from '~/stores/streamSegment'
 const streamSegmentStore = useStreamSegmentStore()
